ACMI-55 AMCI-56 - Fixed recordings from not redirecting property with different group modes are used within the plugin
Better defined the behaviour of how the plugin behaves in group mode and correct the code
to reflect this behaviour.

Fixed the display of meting recording to abide by the correct groups mode.  Now recordings
will only show if they belong to the currently selected group.

Fiuxed recording meetings group mode permissions to not allow users who are not a part of the group
or have permissions from joining the recording